requirements.txt
Flask
mysql-connector-python
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
python-dotenv
web3
mysqlclient